Move setInsetImageDrawable adapter to tract-renderer and disable databinding for tips-renderer
The only remaining usage is the call to action tip arrow in tract_page.xml,
so move the adapter to the tract-renderer module, drop the unused srcInset
overload, and disable databinding (and legacy kapt) for tips-renderer.
